@@ -6,23 +6,30 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 
-namespace PolitiLog
+namespace DiscordWikiLog
 {
     class ChangeNotifier
     {
         private EmbedBuilderHelper _builder;
         private SimpleLogger _logger;
+        private l18n _localization;
+
         private string _webhookUrl;
+        private string _apiUrl;
         private string _wikiUrl;
         private int _queryLimit;
 
-        public ChangeNotifier(SimpleLogger logger, string webhookUrl, string wikiUrl, int queryLimit)
+        public ChangeNotifier(SimpleLogger logger, string webhookUrl, string api, string wiki, int queryLimit, l18n localization)
         {
             _logger = logger;
+            _localization = localization;
+
             _webhookUrl = webhookUrl;
-            _wikiUrl = wikiUrl;
+            _apiUrl = api;
+            _wikiUrl = wiki;
             _queryLimit = queryLimit;
-            _builder = new EmbedBuilderHelper(_wikiUrl, _logger);
+
+            _builder = new EmbedBuilderHelper(_wikiUrl, _logger, localization);
         }
 
         public DateTime SendRevisionSinceLastRevision(DateTime lastChange)
@@ -31,7 +38,7 @@ namespace PolitiLog
 
             SendToWebHook(BuildEmbeds(changes));
 
-            _logger.AddLog(String.Format("{0} new changes to publish", changes.Count));
+            _logger.AddLog(changes.Count + " " + "new changes to publish");
 
             // if no changes since last check, return
             if (changes.Count == 0)
@@ -62,14 +69,14 @@ namespace PolitiLog
                 using (WebClient client = new WebClient())
                 {
                     // Query api and parse json
-                    string queryUrl = String.Format("{0}/w/api.php?action=query&list=recentchanges&rcprop=ids|title|user|comment|timestamp&rclimit={1}&format=json", _wikiUrl, _queryLimit);
+                    string queryUrl = String.Format("{0}?action=query&list=recentchanges&rcprop=ids|title|user|comment|timestamp&rclimit={1}&format=json", _apiUrl, _queryLimit);
                     string json = client.DownloadString(queryUrl);
                     JObject jsonObject = JObject.Parse(json);
 
                     // for each revisions, only keep edit, new and log type of revision
                     foreach (var change in jsonObject["query"]["recentchanges"])
                         if (change.Value<string>("type") == "edit" || change.Value<string>("type") == "new" || change.Value<string>("type") == "log")
-                            if (change.Value<DateTime>("timestamp").ToUniversalTime() > lastChange.ToUniversalTime())
+                            if (change.Value<DateTime>("timestamp").ToUniversalTime() < lastChange.ToUniversalTime())
                                 revisions.Add(new Change(change));
 
                     // sort by date
@@ -90,7 +97,7 @@ namespace PolitiLog
                 try
                 {
                     var client = new DiscordWebhookClient(_webhookUrl);
-                    client.SendMessageAsync(text: "Nouveau changement sur PolitiWiki !", embeds: new[] { embed }).Wait();
+                    client.SendMessageAsync(text: _localization.WebhookMessage, embeds: new[] { embed }).Wait();
                 }
                 catch (Exception e){
                     _logger.AddLog(e.Message);
